Mostly Playable
Game mostly plays itself out now.  You can technically continue adding cards past 3, but it won't actually do anything in terms of damage and what not.  After your damage is dealt in increments, the enemy deals 3 separate hits of damage. After this damage is dealt, the cards are reshuffled into the deck and redealt to the player to start a new turn. Turn now increments as well.